refactor(welcome): extract test-connection probe into coreRpcClient
CodeRabbit feedback (#1301): the inline bearer-token + JSON-RPC envelope
assembly in `Welcome.tsx`'s `handleTestConnection` violates the project
guideline that RPC client calls live in `services/coreRpcClient`, not
in components.

Move the probe behind a new `testCoreRpcConnection(url)` helper and
revert the public export of `getCoreRpcToken` (now only the new helper
needs it).
